requirements.dev.txt
djangorestframework
drf-yasg